refactoring numbering plan (change zoneId from int to long)
git-tfs-id: [http://developmentvlan:8585/tfs/vanrise.collection]$/;C32392

+++ TOneV2/Code/TOneV2/TOne.WhS.BusinessEntity.Business/SaleZone/SaleCodeManager.cs
@@ -71,7 +71,7 @@ namespace TOne.WhS.BusinessEntity.Business
             return dataManager.GetSaleCodesEffectiveAfter(sellingNumberPlanId, countryId, minimumDate);
         }
 
-        public List<SaleCode> GetSaleCodesByZoneIDs(List<int> zoneIds, DateTime effectiveDate)
+        public List<SaleCode> GetSaleCodesByZoneIDs(List<long> zoneIds, DateTime effectiveDate)
         {
             ISaleCodeDataManager dataManager = BEDataManagerFactory.GetDataManager<ISaleCodeDataManager>();
             return dataManager.GetSaleCodesByZoneIDs(zoneIds, effectiveDate);
