+++ TOneV2/Code/TOneV2/TOne.WhS.BusinessEntity.Data.SQL/SaleCodeDataManager.cs
@@ -91,11 +91,11 @@ namespace TOne.WhS.BusinessEntity.Data.SQL
             return GetItemsSP("TOneWhS_BE.sp_SaleCode_GetByCountry", SaleCodeMapper, countryId, effectiveDate);
         }
 
-        public List<SaleCode> GetSaleCodesByZoneIDs(List<int> zoneIds, DateTime effectiveDate)
+        public List<SaleCode> GetSaleCodesByZoneIDs(List<long> zoneIds, DateTime effectiveDate)
         {
             string allZoneIds = null;
             if (zoneIds != null && zoneIds.Count() > 0)
-                allZoneIds = string.Join<int>(",", zoneIds);
+                allZoneIds = string.Join<long>(",", zoneIds);
             return GetItemsSP("TOneWhS_BE.sp_SaleCode_GetByZoneIds", SaleCodeMapper, allZoneIds, effectiveDate);
         }
